# cache: Load before Store