#### cache: fix and improve Get benchmark


#### cache: Load before Store

On x86 an atomic.Load is a simple load, whereas a Store is an atomic
`XCHG` operation which involves a memory barrier. Frequently accessed
blocks are likely to be accessed many times between two sweeps of the
clockpro hand.

On a gceworker:
```
name         old time/op    new time/op    delta
CacheGet-24    20.1ns ± 4%    18.9ns ± 5%  -5.82%  (p=0.002 n=10+10)
```

Fixes #4328